backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
pydantic[email]==2.7.1
pydantic-settings==2.3.0
python-jose[cryptography]==3.3.0
redis[asyncio]==5.0.4
httpx==0.27.0
python-multipart==0.0.9
minio==7.2.7
aiofiles==23.2.1
# OCR
pytesseract==0.3.10
Pillow==10.3.0
pdfplumber==0.11.1
opencv-python-headless==4.9.0.80
numpy==1.26.4
# AI
anthropic==0.26.0
google-cloud-vision==3.7.0
pytest==8.2.1
pytest-asyncio==0.23.7
